Change address type to uint64_t for new backends
Keep uint32_t interface for existing backends

+++ device_backends/xdma/include/XdmaBackend.h
@@ -34,8 +34,8 @@ public:
 
     bool isFunctional() const override;
 
-    void read(uint8_t bar, uint32_t address, int32_t* data, size_t sizeInBytes) override;
-    void write(uint8_t bar, uint32_t address, const int32_t* data, size_t sizeInBytes) override;
+    void read(uint8_t bar, uint64_t address, int32_t* data, size_t sizeInBytes) override;
+    void write(uint8_t bar, uint64_t address, const int32_t* data, size_t sizeInBytes) override;
 
     std::string readDeviceInfo() override;
